Make feature list easier to manage
